rag-service/requirements.txt
fastapi>=0.112.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
pyyaml>=6.0.1
pandas>=2.2.2
tqdm>=4.66.5

pymupdf>=1.24.9
paddleocr>=2.8.1
paddlepaddle>=3.3.1
pillow>=10.4.0
numpy>=1.26.0

pypdf>=4.3.1
python-docx>=1.1.2
jieba>=0.42.1
rank-bm25>=0.2.2
faiss-cpu>=1.8.0
sentence-transformers>=3.0.1
openai>=1.40.0
